\name{armengol}
\Rdversion{1.1}
\alias{armengol}
\docType{data}

\title{
CNV data from Armengol, et al paper
}

\description{

}

\usage{data(armengol)}

\format{
  A data frame with 261 observations containing information about 120 CNVs for three human major groups (CEU, YRI and CHB+JPT)
}

\details{

}

\source{

}

\references{
Armengol et al. Identification of Copy Number Variants Defining Genomic Differences among Major Human Groups. Plos One, 2009, e7230+.
}

\examples{
data(armengol)
}

\keyword{datasets}
