\name{bayesCNVassoc}
\Rdversion{1.1}
\alias{bayesCNVassoc}
\alias{plot.bayesCNVassoc}
\alias{print.bayesCNVassoc}

\title{
Bayesian model to analyze CNV data
}

\description{

}

\usage{
bayesCNVassoc(y, cnvs, QC=0.9, method = "JAGS", n.iter.burn.in = 10000, n.iter = 30000, thin = 50, n.chain = 2, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{}
  \item{cnvs}{}
  \item{QC}{}
  \item{method}{}
  \item{n.iter.burn.in}{}
  \item{n.iter}{}
  \item{thin}{}
  \item{n.chain}{}
  \item{\dots}{}
}


\details{


}


\value{
}

\references{

}


\seealso{

}

\examples{
}

\keyword{utilities}

