\name{bayesSNPassoc}
\Rdversion{1.1}
\alias{bayesSNPassoc}
\alias{plot.bayesSNPassoc}
\alias{print.bayesSNPassoc}

\title{
Bayesian model to analyze SNP data
}

\description{

}

\usage{
bayesSNPassoc <- function (y, snps, annotation, chr, QC = 0.9, min.freq=0.05, method = "JAGS", n.iter.burn.in = 10000, n.iter = 30000, thin = 50, n.chain = 2,  
    ...)
}

\arguments{
  \item{y}{}
  \item{snps}{}
  \item{annotation}{}
  \item{chr}{}
  \item{QC}{}
  \item{min.freq}{}
  \item{method}{}
  \item{n.iter.burn.in}{}
  \item{n.iter}{}
  \item{thin}{}
  \item{n.chain}{}
  \item{\dots}{}
}

\details{

}

\value{
}


\references{
}

\seealso{

}

\examples{
}

\keyword{utilities}

