\name{checkConvergence}
\Rdversion{1.1}
\alias{checkConvergence}

\title{
Function to check convergence of Bayesian model
}

\description{

}

\usage{
checkConvergence(x, type = "Markov chain", parameter = "alpha", ...)
}


\arguments{
  \item{x}{}
  \item{type}{}
  \item{parameter}{}
  \item{\dots}{}
}

\details{}

\value{}

\references{

}


\seealso{

}

\examples{

}

\keyword{utilities}
