\name{makeHeatmap}
\Rdversion{1.1}
\alias{makeHeatmap}

\title{
Creates a Heatmap using coefficients of Bayesian model
}

\description{

}

\usage{
makeHeatmap(x, quantiles = c(0.2, 0.5, 0.6, 0.8), ...)
}


\arguments{
  \item{x}{}
  \item{quantiles}{}
  \item{\dots}{}
}

\details{

}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\references{
%% ~put references to the literature/web site here ~
}



\seealso{

}

\examples{
}

\keyword{utilities}

