\name{sim.data}
\Rdversion{1.1}
\alias{sim.data}
\docType{data}

\title{
Simulated SNP data
}

\description{

}

\usage{data(sim.data)}

\format{
  A data frame with 800 observations on the following 73 variables.  
  .....
}

\details{

}

\source{

}

\references{

}

\examples{
data(sim.data)
}

\keyword{datasets}
